+++ demisto_sdk/commands/common/constants.py
@@ -70,6 +70,8 @@ GENERIC_MODULE = 'genericmodule'
 GENERIC_DEFINITION = 'genericdefinition'
 JOB = 'job'
 
+MARKETPLACE_KEY_PACK_METADATA = 'marketplaces'
+
 
 class FileType(Enum):
     INTEGRATION = 'integration'
@@ -1331,3 +1333,8 @@ UUID_REGEX = r'([\w]{8}-[\w]{4}-[\w]{4}-[\w]{4}-[\w]{8,12})'
 
 class IronBankDockers:
     API_LINK = 'https://repo1.dso.mil/api/v4/projects/dsop%2Fopensource%2Fpalo-alto-networks%2Fdemisto%2F'
+
+
+class MarketplaceVersions(Enum):
+    XSOAR = 'xsoar'
+    MarketplaceV2 = 'marketplacev2'